remove abstraction layer for filesystem operations
