+++ crates/polyfuse/src/lib.rs
@@ -6,9 +6,8 @@ mod write;
 
 pub mod op;
 pub mod reply;
-pub mod request;
 
 pub use crate::{
-    request::{Operation, Request},
-    session::{CapabilityFlags, Config, ConnectionInfo, Session},
+    op::Operation,
+    session::{CapabilityFlags, Config, ConnectionInfo, Request, Session},
 };
